backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-jose
passlib
python-multipart
bcrypt
email-validator
python-dotenv
